+++ .idea/misc.xml
@@ -5,20 +5,23 @@
       <map>
         <entry key="..\:/Github/Android-practice/tarsosDSP_PitchDetect/app/src/main/res/layout/activity_main.xml" value="0.4" />
         <entry key="..\:/Github/PitchDetection/app/src/main/res/layout/activity_main.xml" value="0.23802083333333332" />
-        <entry key="..\:/Github/PitchDetection/app/src/main/res/layout/activity_octave_test.xml" value="0.2857638888888889" />
+        <entry key="..\:/Users/mok33/OneDrive/PitchDetection/app/src/main/res/layout/activity_main.xml" value="0.2203125" />
         <entry key="app/src/main/res/drawable/outline_arrow_back_ios_new_20.xml" value="0.178125" />
+        <entry key="app/src/main/res/drawable/singing_standby_keybutton_background.xml" value="0.2984375" />
         <entry key="app/src/main/res/layout/activity_custom_song_list.xml" value="0.29583333333333334" />
         <entry key="app/src/main/res/layout/activity_email_password.xml" value="0.2942708333333333" />
         <entry key="app/src/main/res/layout/activity_log_in.xml" value="0.28141511226244353" />
         <entry key="app/src/main/res/layout/activity_login.xml" value="0.21300510243705637" />
         <entry key="app/src/main/res/layout/activity_main.xml" value="0.2921935313541035" />
+        <entry key="app/src/main/res/layout/activity_music_is_my_life.xml" value="0.29583333333333334" />
         <entry key="app/src/main/res/layout/activity_register.xml" value="0.3126618054457129" />
+        <entry key="app/src/main/res/layout/activity_singing_standby.xml" value="0.3098958333333333" />
         <entry key="app/src/main/res/layout/activity_song_list.xml" value="0.26222826086956524" />
         <entry key="app/src/main/res/layout/custom_listview.xml" value="0.40913252683940676" />
       </map>
     </option>
   </component>
-  <component name="ProjectRootManager" version="2" languageLevel="JDK_11" default="true" project-jdk-name="Android Studio default JDK" project-jdk-type="JavaSDK">
+  <component name="ProjectRootManager" version="2" languageLevel="JDK_13" default="true" project-jdk-name="Android Studio default JDK" project-jdk-type="JavaSDK">
     <output url="file://$PROJECT_DIR$/build/classes" />
   </component>
   <component name="ProjectType">
