+++ app/src/main/AndroidManifest.xml
@@ -22,6 +22,11 @@
         <activity
             android:name=".SongListActivity"
             android:exported="true">
+            <intent-filter>
+                <action android:name="android.intent.action.MAIN" />
+
+                <category android:name="android.intent.category.LAUNCHER" />
+            </intent-filter>
         </activity>
         <activity
             android:name=".OctaveTestEndActivity"
@@ -41,11 +46,7 @@
         <activity
             android:name=".LogInActivity"
             android:exported="true">
-            <intent-filter>
-                <action android:name="android.intent.action.MAIN" />
 
-                <category android:name="android.intent.category.LAUNCHER" />
-            </intent-filter>
         </activity>
     </application>
