+++ app/src/main/res/layout/activity_song_list.xml
@@ -22,7 +22,7 @@
     <ListView
         android:id="@+id/songListArea"
         android:layout_width="329dp"
-        android:layout_height="501dp"
+        android:layout_height="500dp"
         android:layout_marginStart="40dp"
         android:layout_marginTop="128dp"
         app:layout_constraintStart_toStartOf="parent"
