+++ app/src/main/res/layout/custom_listview.xml
@@ -1,15 +1,17 @@
 <?xml version="1.0" encoding="utf-8"?>
 <LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
     android:layout_width="300dp"
-    android:layout_height="100dp"
+    android:layout_height="wrap_content"
     android:orientation="horizontal"
     android:padding="10dp">
 
     <TextView
         android:id="@+id/indexTextView"
         android:layout_width="20dp"
-        android:layout_height="20dp"
+        android:layout_height="40dp"
         android:layout_gravity="center"
+        android:layout_marginTop="10dp"
         android:layout_marginRight="10dp"
         android:text="1"
         android:textAlignment="center" />
@@ -26,12 +28,24 @@
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:text="title"
-            android:textStyle="bold" />
+            android:textSize="18sp" />
 
         <TextView
             android:id="@+id/singerTextView"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
+            android:textSize="14sp"
             android:text="content" />
     </LinearLayout>
+
+    <ImageButton
+        android:id="@+id/playingButton"
+        android:background="@android:color/transparent"
+        android:layout_width="20dp"
+        android:layout_height="30dp"
+        android:layout_weight="1"
+        android:layout_marginTop="10dp"
+        android:scaleType="centerInside"
+        app:srcCompat="@drawable/play" />
+
 </LinearLayout>
